refactor(server): Move CaseChecker to zally-core #560
